fix(mm): guard compaction/damon/madvise arithmetic against overflow (#81)
Defensive hardening of latent ring-0 panics in exported-but-unwired mm
modules. Each site consumes an order, pfn range, total, or address+len that
would panic under overflow-checks once a live caller is wired; the fixes
mirror the defensive pattern already used by sibling functions.

Compaction (compaction.rs, compaction_scan.rs, compaction_zone.rs):
- should_compact: `1 << order` used checked_shl (saturate) so an order >= 64
  cannot fault; the free-path helper additionally rejects order >= 64.
- the free-path ratio test guards total_pages == 0 before dividing and is
  rearranged to a cross-multiply (saturating_mul) to avoid both the
  divide-by-zero and the product overflow.
- new_zone makes the midpoint computation underflow-safe and add_zone
  rejects an inverted pfn range (start >= end), mirroring Compactor::add_zone.
- scan_pass / progress_pct use saturating arithmetic for the cursor and
  percentage math.

DAMON (damon_core.rs, damon_region.rs):
- DamonRegion::new normalizes an inverted range and size() uses
  saturating_sub; merge_regions clamps the merged end with max so it can
  never fall below start; access_rate_pct computes the *100 in u64.

madvise (madvise_ops.rs):
- do_madvise computes the range end with checked_add and rejects an
  overflowing (start, len).

All signatures and const-ness preserved; every guard fails only on a
genuine overflow / inverted range, and legit orders, ordered ranges, and
realistic counts produce identical results. False positives (already-guarded
damon add_region / stats) were left untouched.

+++ crates/mm/src/damon_region.rs
@@ -132,7 +132,11 @@ impl DamonRegion {
         if self.sample_attempts == 0 {
             return 0;
         }
-        self.access_count * 100 / self.sample_attempts
+        // SECURITY: widen to u64 before the *100 so a large access_count
+        // cannot overflow u32 (panics with overflow-checks on). The ratio is
+        // bounded (<=100 for realistic counts) so the cast back is exact and
+        // matches the previous result for all realistic sample counts.
+        ((self.access_count as u64 * 100) / self.sample_attempts as u64) as u32
     }
 
     /// Record a sample.
